refactor: deletePlaceLike를 v1으로 이동
