+++ backend/spring-routie/src/main/java/routie/business/like/ui/v2/PlaceLikeControllerV2.java
@@ -2,7 +2,6 @@ package routie.business.like.ui.v2;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -25,16 +24,6 @@ public class PlaceLikeControllerV2 {
             @AuthenticatedParticipant final Participant participant
     ) {
         participantPlaceLikeService.likePlace(placeId, routieSpaceIdentifier, participant);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/likes")
-    public ResponseEntity<Void> deleteLike(
-            @PathVariable final Long placeId,
-            @PathVariable final String routieSpaceIdentifier,
-            @AuthenticatedParticipant final Participant participant
-    ) {
-        participantPlaceLikeService.removePlaceLike(placeId, routieSpaceIdentifier, participant);
         return ResponseEntity.ok().build();
     }
 }
