+++ backend/spring-routie/src/test/java/routie/business/like/ui/v2/PlaceLikeControllerV2Test.java
@@ -1,4 +1,4 @@
-package routie.business.like.ui.v1;
+package routie.business.like.ui.v2;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -14,7 +14,11 @@ import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
+import routie.business.authentication.domain.jwt.JwtProcessor;
 import routie.business.like.domain.PlaceLikeRepository;
+import routie.business.participant.domain.User;
+import routie.business.participant.domain.UserFixture;
+import routie.business.participant.domain.UserRepository;
 import routie.business.place.domain.Place;
 import routie.business.place.domain.PlaceBuilder;
 import routie.business.place.domain.PlaceRepository;
@@ -24,7 +28,7 @@ import routie.business.routiespace.domain.RoutieSpaceRepository;
 
 @DirtiesContext(classMode = ClassMode.AFTER_CLASS)
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-public class PlaceLikeControllerV1Test {
+public class PlaceLikeControllerV2Test {
 
     @LocalServerPort
     private int port;
@@ -40,6 +44,12 @@ public class PlaceLikeControllerV1Test {
 
     @Autowired
     private PlaceLikeRepository placeLikeRepository;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private JwtProcessor jwtProcessor;
 
     private Place testPlace;
     private RoutieSpace testRoutieSpace;
@@ -62,16 +72,23 @@ public class PlaceLikeControllerV1Test {
     }
 
     @Test
-    @DisplayName("V1 API로 장소에 좋아요를 성공적으로 추가한다")
+    @DisplayName("V2 API로 장소에 좋아요를 성공적으로 추가한다")
     public void likePlace() {
         // given
         long placeId = testPlace.getId();
         long initialLikeCount = placeLikeRepository.countByPlace(testPlace);
 
+        // 테스트용 사용자 생성 및 토큰 발급
+        User user = UserFixture.emptyUser();
+        userRepository.save(user);
+        String accessToken = jwtProcessor.createJwt(user);
+
         // when
         Response response = RestAssured
+                .given().log().all()
+                .header("Authorization", "Bearer " + accessToken)
                 .when()
-                .post("/v1/routie-spaces/" + testRoutieSpace.getIdentifier() + "/places/" + placeId + "/likes")
+                .post("/v2/routie-spaces/" + testRoutieSpace.getIdentifier() + "/places/" + placeId + "/likes")
                 .then()
                 .log().all()
                 .extract().response();
@@ -87,16 +104,23 @@ public class PlaceLikeControllerV1Test {
     }
 
     @Test
-    @DisplayName("V1 API로 존재하지 않는 루티 스페이스에 좋아요 시 예외가 발생한다")
+    @DisplayName("V2 API로 존재하지 않는 루티 스페이스에 좋아요 시 예외가 발생한다")
     public void likeWithNonExistentRoutieSpace() {
         // given
         long placeId = testPlace.getId();
         String nonExistentIdentifier = "non-existent-identifier";
 
+        // 테스트용 사용자 생성 및 토큰 발급
+        User user = UserFixture.emptyUser();
+        userRepository.save(user);
+        String accessToken = jwtProcessor.createJwt(user);
+
         // when
         Response response = RestAssured
+                .given().log().all()
+                .header("Authorization", "Bearer " + accessToken)
                 .when()
-                .post("/v1/routie-spaces/" + nonExistentIdentifier + "/places/" + placeId + "/likes")
+                .post("/v2/routie-spaces/" + nonExistentIdentifier + "/places/" + placeId + "/likes")
                 .then()
                 .log().all()
                 .extract().response();
@@ -109,15 +133,22 @@ public class PlaceLikeControllerV1Test {
     }
 
     @Test
-    @DisplayName("V1 API로 존재하지 않는 장소에 좋아요 시 예외가 발생한다")
+    @DisplayName("V2 API로 존재하지 않는 장소에 좋아요 시 예외가 발생한다")
     public void likeNonExistentPlace() {
         // given
         long nonExistentPlaceId = 999999L;
 
+        // 테스트용 사용자 생성 및 토큰 발급
+        User user = UserFixture.emptyUser();
+        userRepository.save(user);
+        String accessToken = jwtProcessor.createJwt(user);
+
         // when
         Response response = RestAssured
+                .given().log().all()
+                .header("Authorization", "Bearer " + accessToken)
                 .when()
-                .post("/v1/routie-spaces/" + testRoutieSpace.getIdentifier() + "/places/" + nonExistentPlaceId
+                .post("/v2/routie-spaces/" + testRoutieSpace.getIdentifier() + "/places/" + nonExistentPlaceId
                         + "/likes")
                 .then()
                 .log().all()
